Allow random stages to sample from a subset
Random stage selection allows more diverse sampling and exploration for
agents, though the wide variety of stages in Super Mario Bros. can be
challenging for agents to learn. Specifying a subset of stages to
randomly sample allows the user to hand-select similar levels for agents
to have more diverse observations, while still being focused enough to
learn the mechanics of the environments in question.

For example, several of the castle levels have identical layouts but
feature slightly different hazards. Allowing the agent to learn on
different castle environments can help it be more generic while still
having a reasonable chance of completing the level.

Signed-Off-By: Robert Clark <robdclark@outlook.com>